gchqbot/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord.py==1.3.3
idna==2.9
motor==2.1.0
multidict==4.7.5
pip-autoremove==0.9.1
pymongo==3.10.1
PyYAML==5.3.1
websockets==8.1
yarl==1.4.2
